Moved changes to correponding _extended files
