+++ superset/assets/javascripts/explorev2/stores/controls_extended.jsx
@@ -1,5 +1,6 @@
 import control from './controls';
 import React from 'react';
+import { formatSelectOptionsForRange, formatSelectOptions } from '../../modules/utils';
 
 const newControls = {
     // Pie Chart
@@ -72,8 +73,16 @@ const newControls = {
             description: 'Select column containing Y ticks'
     },
 
-
-
+    //Collapsible_Force Form Parameters
+    gravity: {
+        type: 'SelectControl',
+        freeForm: true,
+        label: 'Gravity',
+        default: '0.3',
+        choices: formatSelectOptions(['0.05','0.1','0.15','0.2','0.25','0.3','0.35','0.4','0.45','0.5','0.55','0.6','0.65',
+            '0.7','0.75','0.8','0.85','0.9','0.95','1.0',]),
+        description: 'Gravity in the force layout',
+    }
 
 }
