+++ superset/assets/visualizations/main.js
@@ -30,9 +30,6 @@ const vizMap = {
   word_cloud: require('./word_cloud.js'),
   world_map: require('./world_map.js'),
   dual_line: require('./nvd3_vis.js'),
-
-  collapsible_force: require('./collapsible_force.js'),
-  coffee_wheel: require('./coffee_wheel.js'),
 };
 
 export default vizMap;
