+++ superset/assets/visualizations/main_extended.js
@@ -4,5 +4,7 @@ vizMap['default']['UkViz'] = require('./uk_vis.js');
 vizMap['default']['LinePlusBarChartViz'] = require('./nvd3_vis_extended.js');
 vizMap['default']['BubbleWithFilterViz'] = require('./nvd3_vis_extended.js');
 vizMap['default']['SunburstIntensityViz'] = require("./sunburst_intensity.js");
+vizMap['default']['collapsible_force'] = require('./collapsible_force.js');
+vizMap['default']['coffee_wheel'] = require('./coffee_wheel.js');
 
 export default vizMap;
